% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fremontBikes.R
\docType{data}
\name{fremontBikes}
\alias{fremontBikes}
\title{Seattle Fremont Bridge Bike Counter.}
\format{A data frame with XXX rows and XXX variables:
\describe{
  \item{date}{Date}
  \item{fremont_bridge_total}{description}
  \item{fremont_bridge_east_sidewalk}{description}
  \item{fremont_bridge_west_sidewalk}{description}
}}
\source{
\url{https://data.seattle.gov/Transportation/Fremont-Bridge-Bicycle-Counter/65db-xm6k}
}
\usage{
fremontBikes
}
\description{
A dataset of XXX.
}
\keyword{datasets}
